refactor: draw tracks onto a different canvas
Projecting the full linestring paths for each track every draw was
causing poor performance. The visualization slowed down over time as
tracks increased in length and sped up again after reset.

Using a separate canvas for tracks means that the tracks canvas does
not need to be cleared every draw step. Instead we can project and
draw the only the track for the last step.